Numeric.Wavelet.Continuous

Continuous Wavelet Transform

newtype CWD v n m a b

Constructors

CWD

Fields

cwdLines :: Vector m (CWDLine v n a b)

mapCWD :: (Vector v b, Vector v c) => (b -> c) -> CWD v n m a b -> CWD v n m a c

data CWDLine v n a b

Constructors

CWDLine

Fields

cwdlData :: Vector v n b

cwdlScale :: Finite ((n `Div` 2) + 1)

Scale factor, in number of ticks.

cwdlFreq :: a

The frequency associated with this scale, in inverse tick

cwdlCoI :: Finite ((n `Div` 2) + 1)

How many items are outside of the Cone of Influence, on each side.

mapCWDLine :: (Vector v b, Vector v c) => (b -> c) -> CWDLine v n a b -> CWDLine v n a c

data CWDOpts n

Constructors

CWDO

Fields

cwdoNorm :: LNorm

wavelet normalization

cwdoMinScale :: Finite ((n `Div` 2) + 1)

min scale (period)

cwdoMaxScale :: Finite ((n `Div` 2) + 1)

max scale (period)

defaultCWDO :: KnownNat n => CWDOpts n

cwd :: forall v n m a b. (Vector v (Complex b), KnownNat n, KnownNat m, FFTWReal b, 1 <= n, RealFloat a) => AWavelet v a (Complex b) -> CWDOpts n -> Vector v n (Complex b) -> CWD v n m a (Complex b)

cwd for complex-valued analytic wavelets.

cwdReal :: forall v n m a b. (Vector v b, Vector v (Complex b), KnownNat n, KnownNat m, FFTWReal b, 1 <= n, RealFloat a) => AWavelet v a b -> CWDOpts n -> Vector v n b -> CWD v n m a b

cwd for real-valued analytic wavelets.

Wavelets

data AWavelet v a b

Analytical Wavelet

Constructors

AW

Fields

awVector :: a -> v b

generate a vector within awRange with a given dt

awFreq :: a

Dominant frequency component

awRange :: a

range away from zero outside of which wavelet can be considered negligible

mapAWavelet :: (Vector v b, Vector v c) => (b -> c) -> AWavelet v a b -> AWavelet v a c

morlet :: (Vector v (Complex a), RealFloat a) => a -> AWavelet v a (Complex a)

morletFunc :: RealFloat a => a -> a -> Complex a

meyer :: (Vector v a, RealFloat a) => AWavelet v a a

meyerFunc :: RealFloat a => a -> a